+++ EcoMarket/Scenes/Home2/Sections/TopSection/ProductModel.swift
@@ -25,14 +25,4 @@ extension ProductModel {
          .init(name: "Bubble Elastic T-shirt", brand: "Roller Rabbit", price: "$39.00", image: "hometheme2/image-5"),
          .init(name: "Vado Odelle Dress", brand: "Madewell", price: "$98.00", image: "hometheme2/image")
     ]
-//    {
-//        get async throws {
-//            [.init(name: "Vado Odelle Dress", brand: "Roller Rabbit", price: "$198.00", image: "products/image 90-2"),
-//             .init(name: "Giselle Top in White Linen", brand: "Madewell", price: "$108.00", image: "products/image 90-3"),
-//             .init(name: "Vado Odelle Dress", brand: "Roller Rabbit", price: "$96.00", image: "products/image 90-4"),
-//             .init(name: "Irregular Rib Skirt", brand: "Madewell", price: "$78.00", image: "products/image 90"),
-//             .init(name: "Bubble Elastic T-shirt", brand: "Roller Rabbit", price: "$39.00", image: "products/image 90-2"),
-//             .init(name: "Vado Odelle Dress", brand: "Madewell", price: "$98.00", image: "products/image 90-3")]
-//        }
-//    }
 }
